optimize the code

+++ startup.py
@@ -38,7 +38,9 @@ def download_file(url):
 
 
 def get_latest_provider_config(provider_url):
+    provider_url = provider_url.strip(' ').strip('"')
     if not provider_url.lower().startswith('http'):
+        logger.error('provider url invalid, url={}'.format(provider_url))
         return None
     res = requests.get(provider_url, verify=False)
     logger.info('provider request status code={}'.format(res.status_code))
@@ -76,7 +78,7 @@ def update_clash_config(provider_config_tmp=None):
             local_config = yaml.safe_load(f)
 
     if proxy_authentication_env is not None:
-        proxy_authentication_list_tmp = proxy_authentication_env.split(" ")
+        proxy_authentication_list_tmp = proxy_authentication_env.strip(' ').strip('"').split(" ")
         proxy_authentication_list = []
         for proxy_authentication_tmp in proxy_authentication_list_tmp:
             if len(proxy_authentication_tmp.split(':')) == 2:
@@ -108,6 +110,7 @@ if __name__ == '__main__':
     cpu_type = platform.machine()
     logger.info("os type={}; cpu type={}".format(os_type, cpu_type))
 
+    provider_check_time = time.time()
     logger.info('start to get latest provider config...')
     provider_config = get_latest_provider_config(provider_subscribe_url)
 
@@ -120,20 +123,25 @@ if __name__ == '__main__':
     time.sleep(check_interval_time)
 
     while True:
-        provider_config_old = provider_config
-        logger.info('start to get latest provider config...')
-        provider_config = get_latest_provider_config(provider_subscribe_url)
-        if provider_config_old != provider_config:
-            logger.info('provider config has updated...')
-            logger.info('start update clash config file ...')
-            update_clash_config(provider_config)
-            logger.info('kill current clash app...')
-            clash_app.kill()
-            time.sleep(3)
-            logger.info('start up clash app...')
-            clash_app = subprocess.Popen('/clash')
-        else:
-            logger.info('provider config not change...')
+
+        if time.time() - provider_check_time > 36000:
+            # get provider config 10 hour onetime
+            provider_config_old = provider_config
+
+            logger.info('start to get latest provider config...')
+            provider_check_time = time.time()
+            provider_config = get_latest_provider_config(provider_subscribe_url)
+            if provider_config_old != provider_config:
+                logger.info('provider config has updated...')
+                logger.info('start update clash config file ...')
+                update_clash_config(provider_config)
+                logger.info('kill current clash app...')
+                clash_app.kill()
+                time.sleep(3)
+                logger.info('start up clash app...')
+                clash_app = subprocess.Popen('/clash')
+            else:
+                logger.info('provider config not change...')
 
         time.sleep(check_interval_time)
